requirements.txt
flask
gunicorn
python-dotenv
flask-wtf
pymongo
flask-security
email_validator
passlib
pymongo[srv]
Flask-Mail==0.9.1
